save button working

+++ album/views/photoPost.py
@@ -1,6 +1,8 @@
 from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib.auth.models import User
 from ..models import Foto
+from django.http import HttpResponseRedirect
+from django.urls import reverse
 
 def publish(req):
     '''Pagina de publicação da foto'''
@@ -20,6 +22,7 @@ def publish(req):
         #    infos['publishData'] = date.today()
             photo = Foto.objects.create(
                 photo       = infos['photo'],
+                photo_512 = infos['photo'],
                 description = infos['description'],
                 profile     = infos['profile'],
                 private     = infos['private'],
@@ -28,6 +31,7 @@ def publish(req):
         else:
             photo = Foto.objects.create(
                 photo       = infos['photo'],
+                photo_512   = infos['photo'],
                 description = infos['description'],
                 profile     = infos['profile'],
                 private     = infos['private'],
@@ -38,6 +42,8 @@ def publish(req):
 
 def edit(req, foto_id):
     '''Página de edição da publicação'''
+    if not req.user.is_authenticated:
+        return redirect("login")
     photo = get_object_or_404(Foto, pk=foto_id)
     if photo.profile != req.user:
         return redirect('index')
@@ -48,6 +54,8 @@ def edit(req, foto_id):
 
 def updatePhoto(req, foto_id):
     '''Atualiza as informações da foto'''
+    if not req.user.is_authenticated:
+        return redirect("login")
     if req.method == 'POST':
         photo = get_object_or_404(Foto, pk=foto_id)
         photo.description = req.POST['description']
@@ -56,7 +64,18 @@ def updatePhoto(req, foto_id):
 
 def deletePhoto(req, foto_id):
     '''Deleta a foto'''
+    if not req.user.is_authenticated:
+        return redirect("login")
     photo = get_object_or_404(Foto, pk=foto_id)
     photo.delete()
     return redirect('index')
 
+def like(req, foto_id):
+    if not req.user.is_authenticated:
+        return redirect("login")
+    post = get_object_or_404(Foto, id=foto_id)
+    if req.user in post.likes.all():
+        post.likes.remove(req.user)
+    else:
+        post.likes.add(req.user)
+    return HttpResponseRedirect(reverse('index'))
